+++ elements/emoji/src/components/typeahead/EmojiTypeAheadComponent.tsx
@@ -19,6 +19,7 @@ import {
   typeaheadCancelledEvent,
   typeaheadSelectedEvent,
   typeaheadRenderedEvent,
+  ufoExperiences,
 } from '../../util/analytics';
 import { EmojiContext } from '../common/internal-types';
 import { createRecordSelectionDefault } from '../common/RecordSelectionDefault';
@@ -148,6 +149,8 @@ export default class EmojiTypeAheadComponent extends PureComponent<
         typeaheadCancelledEvent(Date.now() - this.openTime, query, emojis),
       );
     }
+    ufoExperiences['emoji-searched'].abort();
+    ufoExperiences['emoji-selection-recorded'].abort();
     this.sessionId = uuid();
     this.selected = false;
   }
@@ -221,6 +224,12 @@ export default class EmojiTypeAheadComponent extends PureComponent<
       options.sort = SearchSort.UsageFrequency;
     }
 
+    ufoExperiences['emoji-searched'].start();
+    ufoExperiences['emoji-searched'].addMetadata({
+      queryLength: query?.length || 0,
+      source: 'typeahead',
+    });
+
     this.renderStartTime = Date.now();
 
     emojiProvider.filter(query, options);
@@ -233,6 +242,11 @@ export default class EmojiTypeAheadComponent extends PureComponent<
     this.fireAnalyticsEvent(
       typeaheadRenderedEvent(Date.now() - this.renderStartTime, query, emojis),
     );
+
+    ufoExperiences['emoji-searched'].success({
+      metadata: { emojisLength: emojis.length },
+    });
+
     debug(
       'emoji-typeahead.applyPropChanges',
       emojis.length,