+++ media/smart-card/src/view/FlexibleCard/components/blocks/title-block/types.ts
@@ -1,6 +1,7 @@
 import { BlockProps, ElementItem, ActionItem } from '../types';
 import { SmartLinkPosition, SmartLinkTheme } from '../../../../../constants';
 import { RetryOptions } from '../../../types';
+import { LinkProps } from '../../elements/link/types';
 
 export type TitleBlockProps = BlockProps & {
   maxLines?: number;
@@ -13,4 +14,5 @@ export type TitleBlockProps = BlockProps & {
   text?: string;
   showActionOnHover?: boolean;
   onClick?: React.EventHandler<React.MouseEvent | React.KeyboardEvent>;
+  anchorTarget?: LinkProps['target'];
 };