+++ media/smart-card/src/view/FlexibleCard/components/elements/link/types.ts
@@ -7,4 +7,5 @@ export type LinkProps = ElementProps & {
   text?: string;
   theme?: SmartLinkTheme;
   url?: string;
+  target?: '_blank' | '_self' | '_top' | '_parent';
 };